2024 Olympics Men's Volleyball Analysis
In this group project, my team and I analyzed a dataset of 36,335 rows and 52 variables to develop actionable insights from the 2024 Olympic Games Men's Volleyball tournament to assist in Team USA's ambition to win as hosts in the 2028 Olympic Games. We sorted, cleaned, manipulated, and visualized data and presented findings all using Excel, Tableau, Power BI, Word, and PowerPoint.